housekeeping at quetzal

+++ analyzeInSilicoORSF.m
@@ -27,11 +27,11 @@ switch method
         
     case 1 %2D fitting
         if sflog
-            p = fitGaussOri180(log(RF_insilico.ORSF.sfList), 180/pi*RF_insilico.ORSF.oriList, resp');
+            p = fitGaussOri180(log(double(RF_insilico.ORSF.sfList)), 180/pi*RF_insilico.ORSF.oriList, resp');
             RF_insilico.ORSF.bestSF = exp(p(1));
             RF_insilico.ORSF.sigmaSF = exp(p(4));
         else
-            p = fitGaussOri180(RF_insilico.ORSF.sfList, 180/pi*RF_insilico.ORSF.oriList, resp');
+            p = fitGaussOri180(double(RF_insilico.ORSF.sfList), 180/pi*RF_insilico.ORSF.oriList, resp');
             RF_insilico.ORSF.bestSF = p(1);
             RF_insilico.ORSF.sigmaSF = p(4);
         end
